backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
sqlalchemy>=2.0.36
psycopg2-binary>=2.9.10
python-dotenv>=1.0.1
openai>=1.50.0
supabase>=2.10.0
pydantic>=2.9.0
python-multipart>=0.0.12
